requirements.txt
Django==4.2
djangorestframework==3.14.0
stripe==7.0.0
pillow==9.5.0
gunicorn==20.1.0
whitenoise==6.5.0
